src/backend/requirements.txt
google-cloud-storage
opencv-python-headless
numpy
scipy
google-cloud-firestore
google-cloud-logging
